Refactor dashboard layout structure and reintroduce sidebar functionality
