+++ apps/nextjs/src/app/(dashboard)/dashboard/admin/_components/upload-thing-ssr.tsx
