+++ apps/nextjs/src/app/(dashboard)/dashboard/layout.tsx
@@ -1,35 +1,17 @@
-import { Suspense } from "react";
-
-import {
-  SidebarInset,
-  SidebarProvider,
-} from "@battle-stadium/ui/sidebar";
-
-import type { ChildrenProps } from "~/types";
-import DashboardContentSkeleton from "~/app/(dashboard)/dashboard/_components/content/dashboard-content-skeleton";
-import DashboardHeader from "~/app/(dashboard)/dashboard/_components/content/dashboard-header";
-import { BreadCrumbsProvider } from "~/components/breadcrumbs/context";
-
-interface DashboardLayoutProps extends ChildrenProps {
-  sidebar: React.ReactNode;
+interface LayoutProps {
+  children: React.ReactNode;
 }
 
-export default function DashboardLayout ({ children, sidebar }: Readonly<DashboardLayoutProps>) {
+export default function SharedLayout({ children }: LayoutProps) {
   return (
-    <BreadCrumbsProvider>
-      <SidebarProvider className="border-[1px]">
-        {sidebar}
-        <SidebarInset className="border-l-[1px] bg-transparent">
-          <DashboardHeader />
-          <Suspense fallback={ <DashboardContentSkeleton /> }>
-            { children }
-            <span className="sr-only" role="status" aria-live="polite">
-              Loading dashboard content, please wait...
-            </span>
-          </Suspense>
-        </SidebarInset>
-      </SidebarProvider>
-    </BreadCrumbsProvider>
+    <section
+      role="status"
+      aria-label="Loading dashboard content"
+      className="flex flex-1 flex-col gap-4 p-4"
+    >
+      <div className="min-h-[100dvh] flex-1 rounded-xl bg-neutral-900/50 md:min-h-min">
+        {children}
+      </div>
+    </section>
   );
 }
-
